+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Double-E TinyMCE Changelog
 
+## Version 1.0.2
+Date: 7 April 2026
+
+- Fix: Handle script blocks (e.g., those used by Ninja Forms) when filtering `the_content`.
+
+## Version 1.0.1
+Date: 9 March 2026
+
+- Fix: Separate dev Composer config to fix install issues in Composer-managed sites.
+
 ## Version 1.0.0
 Date: 1 February 2026